chore: renamed to features folder
