+++ app/features/about/card/card.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from "react";
-import styles from "@/app/views/about/card/card.module.css";
+import styles from "features/about/card/card.module.css";
 
 export const Card = ({ children }: PropsWithChildren) => {
   return <div className={styles.card}>{children}</div>;
