+++ app/features/work/work.tsx
@@ -1,5 +1,5 @@
 import styles from "@/app/views/work/work.module.css";
-import { TableContainer } from "./experience/table-container";
+import { TableContainer } from "features/work/experience/table-container";
 
 export const Work = () => {
   return (
